uint to uint256

+++ contracts/IWETH.sol
@@ -8,5 +8,5 @@ interface IWETH is IERC20 {
 
     function deposit() external payable;
 
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
